refactor: set up modal message on submit
