+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="shortcut icon" type="image/jpg" href="./src/favicon.ico"/>
-    <title>week 2</title>
+    <title>week 3</title>
   </head>
   <body>
     <div id="root"></div>
